requirements.txt
apache-airflow==2.7.1
apache-airflow-providers-postgres==5.6.0
apache-airflow-providers-mysql==5.2.1
pandas==2.0.3
psycopg2-binary==2.9.7
mysql-connector-python==8.1.0
python-dotenv==1.0.0 